+++ CHANGELOG.md
@@ -22,7 +22,7 @@ changes.
   + Reference scripts on the layer 2 ledger are non-problematic.
 
  - Introduced a hard-coded limit of 100 ADA for the commit tx on mainnet: help people to not shoot themselves in the foot too hard.
- - Added `ReachedMainnetHardcodedLimit` API error to signal that the hard-coded mainnet ADA limit is reached.
+ - Added `CommittedTooMuchADAForMainnet` API error to signal that the hard-coded mainnet ADA limit is exceeded.
 
 ## [0.9.0] - 2023-03-02